Refactor: 컴포넌트 분리!, treemap 그래프 기능 일부 수정 treemap에서 클릭시 화면 안움직이고, 그래프 줌인 줌아웃 기능 삭제
